requirements.txt
openai==1.13.3
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.0
tqdm==4.65.0